Revert incorrect fix to FileArray type
In #47418 , FileArray's values, which can be either a single UploadedFile object or an array of UploadedFiles, were incorrectly changed to reflect just the single object case. The same incorrect change had been attempted in both #40915 and #45572 , and rejected both times because it did not match the actual behavior of the library. Somehow #47418 got merged, though.